refactor scrollama
